调整index 为 INT32类型，此改动对程序正确性无影响。 调整前为UINT32，存在一些小问题：1、调用_metaFilte传入参数时隐式转换为INT32。2、在与fileCount比较时需要强转fileCount为UINT32
git-svn-id: http://192.168.20.11/sequoiadb/trunk@30524 6121f079-4a18-41d3-8c30-8870a322f2c2

+++ SequoiaDB/engine/tools/sdbDpsDump.cpp
@@ -1011,7 +1011,7 @@ INT32 _dpsDumper::_analysisMeta()
    }
    else
    {
-      UINT32 index = 0 ;
+      INT32 index = 0 ;
       ossMemcpy( dirPath, srcPath, OSS_MAX_PATHSIZE ) ;
       rc = getFileCount( dirPath, fileCount ) ;
       if( SDB_OK != rc )
@@ -1028,7 +1028,7 @@ INT32 _dpsDumper::_analysisMeta()
 
       {
          fs::path fileDir( dirPath ) ;
-         for ( index = 0; index < (UINT32)fileCount; ++index )
+         for ( index = 0; index < fileCount; ++index )
          {
             const CHAR *filepath = fileDir.string().c_str() ;
             CHAR filename[ OSS_MAX_PATHSIZE * 2 ] = { 0 } ;
